error in taking inputs

+++ ReservationServiceConsumer/src/reservationserviceconsumer/Activator.java
@@ -26,50 +26,53 @@ public class Activator implements BundleActivator {
 	
 	public void displayMainMenu(ResturantServicePublish resturantServicePublish) {
 		
+		String type;
 		int option;
 		String suboption = "y";
 		
 		Scanner sc = new Scanner(System.in);
 		System.out.println("\n\n");
 		System.out.println("----------Imeero Family Restaurant - Reservation ----------");
-		System.out.println("1  - Add Reservation");
-		System.out.println("2  - Get all Reservations");
-		System.out.println("3  - Get reservation by name");
-		System.out.println("4  - Delete reservation");
-		System.out.println("5  - Get priority/normal reservations");
-
-		
-		System.out.print("Choose an option : ");
-		
-		option = Integer.parseInt(sc.nextLine().trim());
-		
-		switch(option) {
-			case 1:
-				resturantServicePublish.insertReservation();
-				
-				while(suboption.equals("y")) {
-					System.out.println("\n\nDo you want to Add Another Reservation(y/n)");
-					suboption = sc.nextLine().trim();
-		
-					if(suboption.equals("y")) {
-						resturantServicePublish.insertReservation();
-					}
+		System.out.println(" Enter (c) if you are a customer and (a) if you are an admin:");
+		type = sc.nextLine().trim();
+		if(type.equals("c")||type.equals("C")) {
+			resturantServicePublish.insertReservation();
+			
+			while(suboption.equals("y")) {
+				System.out.println("\n\nDo you want to Add Another Reservation(y/n)");
+				suboption = sc.nextLine().trim();
+	
+				if(suboption.equals("y")) {
+					resturantServicePublish.insertReservation();
 				}
-				displayMainMenu(resturantServicePublish);
-				break;
-			case 2:
+			}
+			displayMainMenu(resturantServicePublish);
+		}
+		else if(type.equals("a")||type.equals("A")) {
+			
+			System.out.println("1  - Get all Reservations");
+			System.out.println("2  - Get reservation by name");
+			System.out.println("3  - Delete reservation");
+			System.out.println("4  - Get priority/normal reservations");
+
+			
+			System.out.print("Choose an option : ");
+			
+			option = Integer.parseInt(sc.nextLine().trim());
+			switch(option) {
+			case 1:
 				resturantServicePublish.getallReservations();
 				displayMainMenu(resturantServicePublish);
 				break;
-			case 3:
+			case 2:
 				resturantServicePublish.getReservationsByCustName();
 				displayMainMenu(resturantServicePublish);
 				break;
-			case 4:
+			case 3:
 				resturantServicePublish.deleteReservation();
 				displayMainMenu(resturantServicePublish);
 				break;
-			case 5:
+			case 4:
 				resturantServicePublish.getPriorityNormal();
 				displayMainMenu(resturantServicePublish);
 				break;
@@ -77,6 +80,13 @@ public class Activator implements BundleActivator {
 				System.out.println("Incorrect Input. Try Again...");
 				displayMainMenu(resturantServicePublish);
 		}
+		}
+		
+		
+		
+		
+		
+		
 		
 		
 	}
